+++ DelvUI/changelog.md
@@ -1,5 +1,6 @@
 # 0.3.0.1
 - Fixed smooth HP in party frames.
+- Fixed BLM's Polyglot Bar Glow not hiding.
 
 # 0.3.0.0
 This version brings A LOT of changes. Chances are your settings will be invalid / broken. We recommend starting from a clean state.